+++ examples/logs-agent-iks/provider.tf
@@ -15,3 +15,19 @@ provider "helm" {
     cluster_ca_certificate = data.ibm_container_cluster_config.cluster_config.ca_certificate
   }
 }
+
+locals {
+  at_endpoint = "https://api.${var.region}.logging.cloud.ibm.com"
+}
+
+provider "logdna" {
+  alias      = "at"
+  servicekey = module.observability_instances.activity_tracker_resource_key != null ? module.observability_instances.activity_tracker_resource_key : ""
+  url        = local.at_endpoint
+}
+
+provider "logdna" {
+  alias      = "ld"
+  servicekey = module.observability_instances.log_analysis_resource_key != null ? module.observability_instances.log_analysis_resource_key : ""
+  url        = local.at_endpoint
+}
