+++ examples/logs-agent-roks/outputs.tf
@@ -1,0 +1,11 @@
+##############################################################################
+# Outputs
+##############################################################################
+
+#output "myoutput" {
+#  description = "Description of my output"
+#  value       = "value"
+#  depends_on  = [<some resource>]
+#}
+
+##############################################################################
