+++ examples/logs-agent-roks/variables.tf
@@ -1,0 +1,47 @@
+variable "ibmcloud_api_key" {
+  type        = string
+  description = "The IBM Cloud api token"
+  sensitive   = true
+}
+
+variable "prefix" {
+  type        = string
+  description = "A prefix for the name of all resources that are created by this example"
+  default     = "direct-to-icl"
+}
+
+variable "resource_group" {
+  type        = string
+  description = "An existing resource group name to use for this example. If not specified, a new resource group is created."
+  default     = null
+}
+
+variable "resource_tags" {
+  type        = list(string)
+  description = "A list of tags to add to the resources that are created."
+  default     = []
+}
+
+variable "access_tags" {
+  type        = list(string)
+  description = "Optional list of access management tags to add to resources that are created"
+  default     = []
+}
+
+variable "region" {
+  type        = string
+  description = "The region where the resources are created."
+  default     = "au-syd"
+}
+
+variable "ocp_version" {
+  type        = string
+  description = "Version of the OCP cluster to provision"
+  default     = null
+}
+
+variable "ocp_entitlement" {
+  type        = string
+  description = "Value that is applied to the entitlements for OCP cluster provisioning"
+  default     = null
+}
