+++ version.tf
@@ -10,7 +10,7 @@ terraform {
     }
     helm = {
       source  = "hashicorp/helm"
-      version = ">= 2.10.0, <3.0.0"
+      version = ">= 2.11.0, <3.0.0"
     }
   }
 }
